+++ api/python/src/pyIterator.hpp
@@ -18,8 +18,8 @@
 
 #include "nanobind/nanobind.h"
 #include "nanobind/make_iterator.h"
-#include <iostream>
-//#include "nanobind/nb_lib.h"
+
+#include <string>
 
 namespace nb = nanobind;