+++ api/rust/include/LIEF/rust/MachO/Binary.hpp
@@ -49,6 +49,7 @@
 
 #include "LIEF/rust/Abstract/Binary.hpp"
 
+#include "LIEF/rust/ObjC/Metadata.hpp"
 
 class MachO_Binary : public AbstractBinary {
   public:
@@ -217,6 +218,14 @@ class MachO_Binary : public AbstractBinary {
 
   auto version_min() const {
     return details::try_unique<MachO_VersionMin>(impl().version_min());
+  }
+
+  auto support_arm64_ptr_auth() const {
+    return impl().support_arm64_ptr_auth();
+  }
+
+  auto objc_metadata() const {
+    return details::try_unique<ObjC_Metadata>(impl().objc_metadata());
   }
 
   static bool is_exported(const MachO_Symbol& symbol) {