+++ doc/sphinx/changelog.rst
@@ -11,6 +11,10 @@ Changelog
   * First (beta) release of the bindings (c.f. :ref:`lief_rust_bindings`)
 
 :ELF:
+  * Add :meth:`lief.ELF.Binary.get_relocated_dynamic_array` which allows
+    to get a **relocated** view of the of init/fini entries. This function can
+    handy ELF init array/fini array functions are defined through relocations.
+    See: :issue:`1058`, :issue:`626`
   * Add support for QNX Stack note (:issue:`1065`)
   * The ``static_symbols`` API functions has been renamed in ``symtab_symbols``.
 
@@ -65,6 +69,9 @@ Changelog
 
 :MachO:
 
+  * Add support to modify Mach-O rpath (see: :issue:`1074`)
+  * Add helper :attr:`lief.MachO.Binary.support_arm64_ptr_auth` to check if a
+    Mach-O binary is supporting ARM64 pointer authentication (arm64e)
   * Fix **major performance issue when processing Mach-O binaries on Windows & macOS**
   * Add generic :class:`lief.MachO.UnknownCommand` to support Apple private Load
     commands not officially supported by LIEF.